int to long

+++ Core/IDay.cs
@@ -2,6 +2,6 @@
 
 public interface IDay
 {
-    public int Part1(FileStream fileStream);
-    public int Part2(FileStream fileStream);
+    public long Part1(FileStream fileStream);
+    public long Part2(FileStream fileStream);
 }
